Replace anonymous from intro
Removing anonymous reference as discussed on Discord. 'Anonymous' is a claim we logically can't make in spite of best effort and current insights; we are not delivering proof. In addition there is the threat model of a phone ending in the wrong hands with the collected data, albeit temporarily stored locally, falling in the wrong hands.

We do want to communicate that the approach is "as privacy preserving as we could possibly make it without it being useless"

anonymous > privacy-preserving
over Waku > over Status' peer-to-peer network (Waku)